---Version 1.3.2--- ------------------- *** version 1.3.2 *** Added an updated manual with thanks to Annette Janssen
*** version 1.3.1 ***
Added the files used in the paper of [Chang et al 2020, Water] to the DATM folder

*** version 1.3.0 ***
--BOYUANT AND N-FIXING BLUEGREEN PHYTOPLANKTON--
-This version of PCLake+ includes the expansion of blue-green algae traits described in Chang et al., 2020 (accepted: Water). In short Bluegreen algae can now be given N-fixing capacity, allowing them to take up N in the form of N2. Also, Bluegreen algae can be given the capacity to become boyuant and create a floating layer at the surface (sDBlueSurf).
For more information we refer to Chang et al., 2020 (Water, in press).
Note: By default these traits are turned off.

--ERROR CATCHING--
-All states are now equiped with statements to stop them from going negative due to integration routine errors. If states become negative they are instead put to a NearZero value prior to being used further on in the model calculations.
Note: No error messages are given when this happens.

--DEPTH UPDATES--
-Depth has been reverted to a single sDepthW state, allowing for easier linkage with other models
-Epi and hypolimnion minimal depth issues have been fixed (could be encountered when flexible water depths were used)
-Added a way to force the system to follow a measured water depth as a time series, adjusting the water inflow/outflow accordingly (mDepthW parameter).

--N LOAD--
-measured time series for N load in the form of Phytoplankton can now be assigned to the model as a parameter
-This update fixes an N mass balance error that could occur when specifying the different fractions of nitrogen in the incoming load directly.

--NON SOURCE CODE UPDATES--
-Cleaned up descriptions
-Updated short names for DELWAQ/DUFLOW
-Small fix to the DATM exporter not exporting the last formula column in the derivative sheet

*** version 1.2.2 ***
Fixed issues with DATM import

*** version 1.2.1 ***
Fixed issues with bifurcations in PCLake+

***Version 1.2.0***
This new version of PCModel has been adjusted to allow all PCModel variants (PCLake, PCDitch, PCLake+) to be compiled seperately so that users will not need to recompile manually when they switch between PCModels.
PCLake+ has had fixes to the transport from Marsh zone to Lake zone (and vice versa), fixing an existing error.
PCLake+ has now been given a path variable in the Control Sheet pointing to the location of the gcc compiler.
PCDitch+ has now been given a path variable in the Control Sheet pointing to the location of the gcc compiler.
PCLake+ is now equipped with a DATM import/export, allowing the user to quickly exchange model configurations and results. Furthermore, this allows for users to include model setup easily and in a transferable manner with their papers and reports.
PCLake+ has had some minor fixes on fish stoichiometry that could lead to mass balance errors.
PCLake+ has been extended with the potential for dynamic plant height, by default this is turned off and users are asked to be mindful that it the model is not calibrated to use it.

***Version 1.1***
This version now PCLake+, which extends the PCLake model to cover a wide range of freshwater lakes that differ in stratification
regime and climate-related processes. To this end, the model was extended with a hypolimnion layer that can be
invoked and configured by forcing functions or by simple built-in empirical relationships that impose stratification.
Further adjustments to the original PCLake model have been made with respect to the calculation of 1)
light irradiation in the water column, 2) evaporation processes and 3) phenology of macrophytes.
See the paper associated with the model for more information: https://www.sciencedirect.com/science/article/pii/S0304380019300201

***Version 1.0***
PCLake, PCDitch DATM versions including user manuals and executables as previously available from the NIOO ftp download location.

Change log:
-PCDitch was rebuild on the PCLake DATM to fix issues with bifurcation analysis

Signed-off-by: pcmodel <pcmodel@nioo.knaw.nl>